Implement a disk scanning service for bootloader behaviour
Begin implementation of an example bootloader that makes use of the
bootloader service to scan disks on boot and when they are attached in
order to discover typical boot configuration (currently just EFI
equivalent).

The purpose of this service and the example are to provide LinuxBoot
style functionality to be used to boot a Linux distribution.